+++ CHANGELOG.md
@@ -43,6 +43,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 - Check group max allowed user when promoting a user.
 - Correct Lua runtime decoding of stream identifying parameters.
 - Correctly use optional parameters when they are passed to group creation operations.
+- Lua runtime operations now observe context cancellation while waiting for an available Lua instance.
 
 ## [2.2.1] - 2018-11-20
 ### Added